Assignment3/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
google-cloud-aiplatform==1.36.4 